+++ engine/src/main/java/dev/gamekit/core/Application.java
@@ -25,6 +25,7 @@ import java.util.List;
 @SuppressWarnings({ "BusyWait", "SynchronizeOnNonFinalField" })
 public abstract class Application {
   public static final long FRAME_TIME_MS = 1000 / 240;
+  public static final long RENDER_TIME_MS = 1000 / 60;
   private static Application instance;
 
   protected final Logger logger = LogManager.getLogger(getClass());
@@ -35,6 +36,7 @@ public abstract class Application {
   private final List<Timeout> newTimeouts;
   private final List<Animation> animations;
   private final WorkerThread audioThread;
+  private final WorkerThread drawThread;
   private boolean isRunning;
   private Scene currentScene;
   private Scene nextScene;
@@ -44,17 +46,17 @@ public abstract class Application {
   }
 
   public Application(Settings settings) {
-    Application.instance = this;
-
     logger.debug("Created application");
     logger.debug(settings);
 
+    Application.instance = this;
     this.settings = settings;
     this.window = new Window();
     this.timeouts = new ArrayList<>();
     this.newTimeouts = new ArrayList<>();
     this.animations = new ArrayList<>();
     this.audioThread = new WorkerThread("audio", FRAME_TIME_MS, Audio::update);
+    this.drawThread = new WorkerThread("draw", RENDER_TIME_MS, this::draw);
     this.isRunning = true;
   }
 
@@ -66,7 +68,7 @@ public abstract class Application {
   public void loadScene(Scene scene) {
     if (scene == null) {
       logger.fatal("Unable to load a null scene");
-      throw new NullPointerException("Unable to load a null scene");
+      throw new IllegalArgumentException("Unable to load a null scene");
     }
 
     logger.debug("Loaded scene: {}", scene.name);
@@ -85,7 +87,8 @@ public abstract class Application {
    */
   public void scheduleTask(Task task, long timeoutMs) {
     if (timeoutMs < 0)
-      throw new RuntimeException("Timeout cannot be negative");
+      throw new IllegalArgumentException("Timeout cannot be negative");
+
     newTimeouts.add(new Timeout(timeoutMs, task));
   }
 
@@ -162,6 +165,7 @@ public abstract class Application {
 
     window.getFrame().setVisible(true);
     audioThread.start();
+    drawThread.start();
   }
 
   /** Called in each frame to update the current scene */
@@ -169,27 +173,34 @@ public abstract class Application {
     animations.forEach(Animation::update);
     timeouts.forEach(Timeout::update);
 
-    if (currentScene != null)
+    if (currentScene != null) {
       currentScene._update();
+    }
 
     animations.removeIf(Animation::isEnded);
     timeouts.removeIf(Timeout::isCompleted);
   }
 
-  /**
-   * Applies the camera's transformation on the {@link Window} scene buffer and renders the
-   * current scene
-   */
+  /** Called in each frame to render the current scene */
   private void render() {
+    if (currentScene != null)
+      currentScene._render();
+  }
+
+  /**
+   * Applies the camera's transformation on the {@link Window} scene buffer and draws the current
+   * scene to the {@link Window}
+   */
+  private void draw() {
     if (currentScene != null) {
-      Camera.update();
+      Camera.apply();
 
       synchronized (currentScene) {
-        currentScene._render();
+        currentScene._draw(window.getDisplayGraphics());
       }
-    }
 
-    window.render();
+      window.draw();
+    }
   }
 
   /** Runs cleanup code at the end of a frame */
@@ -213,8 +224,6 @@ public abstract class Application {
       currentScene = nextScene;
       currentScene._start();
       nextScene = null;
-
-      window.createRenderBuffers();
     }
   }
 
@@ -227,6 +236,9 @@ public abstract class Application {
 
     audioThread.interrupt();
     audioThread.join(1000);
+
+    drawThread.interrupt();
+    drawThread.join(1000);
 
     Audio.dispose();
     IO.dispose();