+++ samples/src/main/java/Demo3Input.java
@@ -17,7 +17,7 @@ import static dev.gamekit.utils.Math.cycle;
  * following actions:
  * <ul>
  *   <li>Creates an {@link Application application}</li>
- *   <li>Overrides the {@link Scene#render()}} method to draw</li>
+ *   <li>Overrides the {@link Scene#render(Renderer)}} method to draw</li>
  *   <li>Detects mouse input using {@link Input} to change the color</li>
  *   <li>Uses the {@link Renderer} to draw box with the current color</li>
  * </ul>
@@ -44,19 +44,17 @@ public class Demo3Input extends Scene {
 
   @Override
   protected void update() {
-    if (Input.isButtonDown(Input.BUTTON_LMB)) {
+    if (Input.isButtonDown(Input.BUTTON_LMB) || Input.isKeyPressed(Input.KEY_SPACE)) {
       colorIndex = cycle(colorIndex + 1, 0, COLORS.length - 1);
     }
   }
 
   @Override
-  public void render() {
+  public void render(Renderer renderer) {
     // Clear the screen with dark gray
-    Renderer.setColor(Color.DARK_GRAY);
-    Renderer.clear();
+    renderer.clear(Color.BLACK);
     // Draw a red box
-    Renderer.setColor(COLORS[colorIndex]);
-    Renderer.fillRect(0, 0, 100, 100);
+    renderer.fillRect(0, 0, 100, 100).withColor(COLORS[colorIndex]);
   }
 
   @Override
@@ -67,7 +65,7 @@ public class Demo3Input extends Scene {
         Padding.options().padding(48),
         Text.create(
           Text.options().alignment(Alignment.CENTER),
-          "Left Mouse Button to change color"
+          "Click the Left Mouse Button or press the Space Bar to change color"
         )
       )
     );